moved game into its own module
