+++ tuning/src/bin/tuner/game/mod.rs
@@ -1,36 +1,19 @@
+mod config;
+
+pub use config::GameConfig;
+
 use std::collections::HashMap;
-use std::fmt;
-use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
-use std::sync::{Arc, Mutex};
 
 use ahash::AHashSet;
-use config::EngineConfig;
 use cozy_chess::{Board, Color, Move};
-use rayon::ThreadPoolBuilder;
 use search::Engine;
 use uci::commands::GoParams;
-use utils::{Book, has_check, has_insufficient_material, has_legal_moves};
-
-use crate::progress::MatchProgress;
+use utils::{has_check, has_insufficient_material, has_legal_moves};
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub enum Outcome {
     Win(Color),
     Draw,
-}
-
-/// Configuration for a game.
-#[derive(Clone, Copy)]
-pub struct GameConfig {
-    pub nodes: u64,
-    pub max_plies: u64,
-
-    // Adjudication
-    pub resign_score: i16,
-    pub resign_moves: u64,
-    pub draw_score: i16,
-    pub draw_moves: u64,
-    pub draw_after: u64,
 }
 
 /// A game between two engines.
@@ -110,7 +93,7 @@ impl Game {
     /// Check if the game could end early as a win or draw.
     ///
     /// Saves time by aborting a game if it is clearly won or drawn.
-    /// Very much inspired by the fastgchess implementation.
+    /// Very much inspired by the fastchess implementation.
     fn adjudicate(&mut self, score: i16, stm: Color) -> Option<Outcome> {
         if score.abs() >= self.config.resign_score {
             self.resign_streak += 1;
@@ -182,116 +165,4 @@ impl Game {
             .filter(|&hash| hash != current)
             .collect()
     }
-}
-
-/// Match result from engine A's point of view.
-#[derive(Clone, Copy, Debug, Default)]
-pub struct Score {
-    pub wins: usize,
-    pub losses: usize,
-    pub draws: usize,
-}
-
-impl Score {
-    fn record(&mut self, outcome: Outcome, perspective: Color) {
-        match outcome {
-            Outcome::Draw => self.draws += 1,
-            Outcome::Win(winner) if winner == perspective => self.wins += 1,
-            Outcome::Win(_) => self.losses += 1,
-        }
-    }
-
-    pub fn played(&self) -> usize {
-        self.wins + self.losses + self.draws
-    }
-
-    pub fn points(&self) -> f64 {
-        self.wins as f64 + 0.5 * self.draws as f64
-    }
-}
-
-impl fmt::Display for Score {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let ratio = self.points() / self.played() as f64;
-
-        write!(
-            f,
-            "{} - {} - {}  [{:.3}]",
-            self.wins, self.losses, self.draws, ratio
-        )
-    }
-}
-
-/// A match between two engine configurations.
-pub struct Match {
-    workers: u64,
-    pairs: u64,
-    game: GameConfig,
-}
-
-impl Match {
-    pub fn new(workers: u64, pairs: u64, game: GameConfig) -> Self {
-        Self {
-            workers,
-            pairs,
-            game,
-        }
-    }
-
-    pub fn play(&self, config_a: &EngineConfig, config_b: &EngineConfig, book: &Book) -> Score {
-        let workers = self.workers.min(self.pairs) as usize;
-        let next = AtomicUsize::new(0);
-        let score = Mutex::new(Score::default());
-        let progress = MatchProgress::new(self.pairs as usize * 2);
-
-        let pool = ThreadPoolBuilder::new()
-            .num_threads(workers)
-            .build()
-            .expect("failed to build match thread pool");
-
-        // Runs the following scope once for each thread in the pool.
-        // This is nice so we only set up engines a and b once per thread per match.
-        pool.broadcast(|_| {
-            let mut a = Engine::new(config_a, Arc::new(AtomicBool::new(false)), load_nnue);
-            let mut b = Engine::new(config_b, Arc::new(AtomicBool::new(false)), load_nnue);
-
-            while (next.fetch_add(1, Ordering::Relaxed) as u64) < self.pairs {
-                let opening = book.random_position();
-
-                // A as white
-                let mut game = Game::new(opening.clone(), self.game);
-                let outcome = game.start(&mut a, &mut b);
-                {
-                    let mut score = score.lock().unwrap();
-                    score.record(outcome, Color::White);
-                    progress.update(&score);
-                }
-
-                // A as black
-                let mut game = Game::new(opening, self.game);
-                let outcome = game.start(&mut b, &mut a);
-                {
-                    let mut score = score.lock().unwrap();
-                    score.record(outcome, Color::Black);
-                    progress.update(&score);
-                }
-            }
-        });
-
-        let score = score.into_inner().unwrap();
-        progress.finish(&score);
-
-        score
-    }
-}
-
-fn load_nnue() -> nnue::Evaluator {
-    // TODO: Consider sharing the model path everywhere
-    const MODEL_PATH: &str = "nnue/model.safetensors";
-
-    let mut varmap = candle_nn::VarMap::new();
-    let mut evaluator = nnue::Evaluator::new(&varmap, &candle_core::Device::Cpu);
-    varmap.load(MODEL_PATH).unwrap();
-    evaluator.enable_nnue();
-    evaluator
 }
